% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bigram_frequencies}
\alias{bigram_frequencies}
\title{Bigram frequencies}
\format{
A data frame with 676 rows and 4 variables:
\describe{
\item{\code{bigram_1}}{character. The first letter of the bigram.}
\item{\code{bigram_2}}{character. The second letter of the bigram.}
\item{\code{count}}{double. Overall count of this bigram within corpus.}
\item{\code{percentage}}{double. Overall percentage of this bigram within corpus.}
}
}
\source{
\url{https://www.norvig.com/mayzner.html}
}
\usage{
bigram_frequencies
}
\description{
The count and proportion of all 676 bigrams (pairs of consecutive letters)
within Norvig's corpus.
}
\keyword{datasets}
