% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported.R
\name{choose_a_letter}
\alias{choose_a_letter}
\title{Choose a letter based on frequency (or frequency at position n within a word)}
\usage{
choose_a_letter(n)
}
\arguments{
\item{n}{If n is a non-zero integer between -7 and 7 (position in word, from 1
(first) to 7 (seventh) or from -1 (last) to -7 (seventh from last)), it
returns a single letter based on letter frequencies at that position only.}
}
\value{
A lower-case letter (character string of \code{nchar} 1) between a and z.
}
\description{
Choose a letter based on frequency (or frequency at position n within a word)
}
\examples{
choose_a_letter(3)
}
