% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{letter_frequencies_by_position}
\alias{letter_frequencies_by_position}
\title{Letter counts/frequencies by position}
\format{
A data frame with 364 rows and 4 variables:
\describe{
\item{\code{position}}{integer. Letter's position within a word.}
\item{\code{letter}}{character. The 26 letters of the English alphabet (lower case).}
\item{\code{count}}{double. Overall count of this letter within corpus, at this position within a word.}
\item{\code{percentage}}{double. Overall percentage of this letter within corpus, at this position within a word.}
}
}
\source{
\url{https://www.norvig.com/mayzner.html}
}
\usage{
letter_frequencies_by_position
}
\description{
From Norvig: "Now we show the letter frequencies by position within word.
That is, the frequencies for just the first letter in each word, just the
second letter, and so on.
We also show frequencies for positions relative to the end of the word:
"-1" means the last letter, "-2" means the second to last, and so on."
}
\keyword{datasets}
