% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{letter_frequencies_overall}
\alias{letter_frequencies_overall}
\title{Letter counts/frequencies overall}
\format{
A data frame with 26 rows and 3 variables:
\describe{
\item{\code{letter}}{character. The 26 letters of the English alphabet (lower case).}
\item{\code{count}}{double. Overall count of this letter within corpus.}
\item{\code{percentage}}{double. Overall percentage of this letter within corpus.}
}
}
\source{
\url{https://www.norvig.com/mayzner.html}
}
\usage{
letter_frequencies_overall
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
