% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sample_letters}
\alias{sample_letters}
\title{Pull a weighted sample of n letters from the alphabet}
\usage{
sample_letters(n)
}
\arguments{
\item{n}{The number of letters to return.}
}
\value{
A vector of lower-case letters.
}
\description{
The \code{replace} parameter is fixed to TRUE, as this is what makes sense given
the frequency-dependent nature of this particular sampling approach.
This function returns a single letter based on a weighted sampling from all
26 letters, based on their overall frequency in Norvig's corpus.
}
\examples{
sample_letters(3)
}
