% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{word_length_frequencies}
\alias{word_length_frequencies}
\title{Word length frequencies in English, from Peter Norvig's analysis}
\format{
A data frame with 23 rows and 3 variables:
\describe{
\item{\code{word_length}}{double. Word length.}
\item{\code{count_millions}}{double. Count of words of this length within corpus (in millions).}
\item{\code{percentage}}{double. Percentage of words of this length within corpus.}
}
}
\source{
\url{https://www.norvig.com/mayzner.html}
}
\usage{
word_length_frequencies
}
\description{
From Norvig:
"Here is the breakdown of mentions (in millions) by word length
(looking like a Poisson distribution).
}
\keyword{datasets}
