% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{write_a_poem}
\alias{write_a_poem}
\title{Writes a poem of pure gibberish}
\usage{
write_a_poem(lines, mean_line_length = 7, cat = TRUE, ...)
}
\arguments{
\item{lines}{The number of lines per stanza. A single integer returns a
single stanza of this many lines. A vector of multiple integers, of length
\code{n}, will return a poem of n stanzas, with lengths as given in the vector.}

\item{mean_line_length}{Line lengths will be generated at random from a
normal distribution around this mean, with SD equal to 1 by default.}

\item{cat}{Boolean. Whether to spew the poem straight to stdout via \code{cat()}
(TRUE, default), or return it invisibly (you'll want to pipe it to an
object or some other function, presumably).}

\item{...}{A place to pass on parameters such as \code{cutoff}, which affects the
available frequencies of the bigrams used to build words.}
}
\value{
A beautiful poem
}
\description{
Writes a poem of pure gibberish
}
\examples{
write_a_poem(c(4, 4), cutoff = 0.01)
}
