% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lettern-package.R
\docType{package}
\name{lettern-package}
\alias{lettern}
\alias{lettern-package}
\title{lettern: Makes Nonsense Words Based on English Letter Frequency Data}
\description{
This constructs "words" based on weighted sampling from letter and ngram frequency data in English, as summarised by Peter Norvig.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/francisbarton/lettern}
  \item Report bugs at \url{https://github.com/francisbarton/lettern/issues}
}

}
\author{
\strong{Maintainer}: Fran Barton \email{fbarton@alwaysdata.net} (\href{https://orcid.org/0000-0002-5650-1176}{ORCID})

}
\keyword{internal}
